create home component, rearrange all state from app to home, add links in nav to home
