+++ src/components/Monster.js
@@ -3,7 +3,11 @@ import React from 'react'
 function Monster() {
     return (
         <div>
-            
+            <h1>Name</h1>
+            <p>Type</p>
+            <p>Hit Points</p>
+            <p>Challenge Rating</p>
+            <p>XP</p>
         </div>
     )
 }
